[department-of-veterans-affairs/va-virtual-agent#59] Extracted Web Chat component
